fix: use sqlite as the only possible database backend

+++ src/api/db.py
@@ -2,7 +2,7 @@ import os
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import DeclarativeBase, sessionmaker
-from sqlalchemy.types import BigInteger
+from sqlalchemy.types import Integer
 
 SQLALCHEMY_DATABASE_URL = os.environ.get(
     "SQLALCHEMY_DATABASE_URL", "sqlite:///./arkheon.db"
@@ -17,7 +17,7 @@ SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 class Base(DeclarativeBase):
     type_annotation_map = {
-        int: BigInteger,
+        int: Integer,
     }
